Adds evaluator authentication flow
Implements a new authentication flow for evaluators,
redirecting them to the evaluation form based on their status.

Adds a middleware to check for evaluator status and redirects
if the evaluator has already completed the evaluation or is not found.

Also displays badge indicator showing if an evaluator has completed their evaluation.